refactor: (#75) presentation module package
