+++ simtong-presentation/src/main/kotlin/team/comit/simtong/common/WebCommonAdapter.kt
@@ -1,10 +1,10 @@
-package team.comit.simtong.domain.common
+package team.comit.simtong.common
 
 import org.springframework.web.bind.annotation.*
 import team.comit.simtong.domain.auth.usecase.ReissueTokenUseCase
 import team.comit.simtong.domain.auth.usecase.dto.TokenResponse
-import team.comit.simtong.domain.common.dto.request.WebFindEmployeeNumberRequest
-import team.comit.simtong.domain.common.dto.response.WebFindEmployeeNumberResponse
+import team.comit.simtong.common.dto.request.WebFindEmployeeNumberRequest
+import team.comit.simtong.common.dto.response.WebFindEmployeeNumberResponse
 import team.comit.simtong.domain.user.usecase.FindEmployeeNumberUseCase
 import team.comit.simtong.domain.user.usecase.dto.FindEmployeeNumberRequest
 import javax.validation.Valid
