+++ simtong-presentation/src/main/kotlin/team/comit/simtong/user/WebUserAdapter.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.user
+package team.comit.simtong.user
 
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.PostMapping
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import team.comit.simtong.domain.auth.usecase.dto.TokenResponse
-import team.comit.simtong.domain.user.dto.request.WebSignInRequest
-import team.comit.simtong.domain.user.dto.request.WebSignUpRequest
+import team.comit.simtong.user.dto.request.WebSignInRequest
+import team.comit.simtong.user.dto.request.WebSignUpRequest
 import team.comit.simtong.domain.user.usecase.SignInUseCase
 import team.comit.simtong.domain.user.usecase.SignUpUseCase
 import team.comit.simtong.domain.user.usecase.dto.SignInRequest
