+++ simtong-presentation/src/main/kotlin/team/comit/simtong/user/dto/request/WebSignInRequest.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.user.dto.request
+package team.comit.simtong.user.dto.request
 
 import javax.validation.constraints.NotBlank
 import javax.validation.constraints.NotEmpty
