+++ simtong-presentation/src/main/kotlin/team/comit/simtong/user/dto/request/WebSignUpRequest.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.user.dto.request
+package team.comit.simtong.user.dto.request
 
 import org.hibernate.validator.constraints.Length
 import javax.validation.constraints.*
